DaynaPort emulation supports NetBSD for MacOS, more unit tests, refactoring

+++ cpp/test/protobuf_util_test.cpp
@@ -193,7 +193,7 @@ TEST(ProtobufUtil, DeserializeMessage)
     auto [fd1, filename1] = OpenTempFile();
     // Data size -1
     buf = { byte { 0xff }, byte { 0xff }, byte { 0xff }, byte { 0xff } };
-    EXPECT_EQ(buf.size(), write(fd1, buf.data(), buf.size()));
+    EXPECT_EQ((ssize_t )buf.size(), write(fd1, buf.data(), buf.size()));
     close(fd1);
     fd1 = open(filename1.c_str(), O_RDONLY);
     ASSERT_NE(-1, fd1);
@@ -203,7 +203,7 @@ TEST(ProtobufUtil, DeserializeMessage)
     auto [fd2, filename2] = OpenTempFile();
     // Data size 2
     buf = { byte { 0x02 }, byte { 0x00 }, byte { 0x00 }, byte { 0x00 } };
-    EXPECT_EQ(buf.size(), write(fd2, buf.data(), buf.size()));
+    EXPECT_EQ((ssize_t )buf.size(), write(fd2, buf.data(), buf.size()));
     close(fd2);
     fd2 = open(filename2.c_str(), O_RDONLY);
     EXPECT_NE(-1, fd2);
@@ -238,13 +238,13 @@ TEST(ProtobufUtil, ReadBytes)
 
     int fd = open("/dev/null", O_RDONLY);
     ASSERT_NE(-1, fd);
-    EXPECT_EQ(0, ReadBytes(fd, buf1));
-    EXPECT_EQ(0, ReadBytes(fd, buf2));
+    EXPECT_EQ(0U, ReadBytes(fd, buf1));
+    EXPECT_EQ(0U, ReadBytes(fd, buf2));
     close(fd);
 
     fd = open("/dev/zero", O_RDONLY);
     ASSERT_NE(-1, fd);
-    EXPECT_EQ(1, ReadBytes(fd, buf1));
-    EXPECT_EQ(0, ReadBytes(fd, buf2));
+    EXPECT_EQ(1U, ReadBytes(fd, buf1));
+    EXPECT_EQ(0U, ReadBytes(fd, buf2));
     close(fd);
 }
